refactor(card): Move all components to one folder
