+++ src/app/components/redir/redir.component.ts
